C11-159: extract workspace.* and pane.* socket handlers
Verbatim relocation into WorkspaceHandlers.swift (25 cases/methods) and
PaneHandlers.swift (13 cases + pane helpers). Router wired by prefix. Files
added to c11 target only. TerminalController.swift: 17959 -> 15589 LOC.
