+++ Sources/TerminalController.swift
@@ -2813,58 +2813,8 @@ class TerminalController {
         // Windows — see Sources/SocketHandlers/WindowHandlers.swift (C11-159)
 
         // Workspaces
-        case "workspace.list":
-            return v2Result(id: id, self.v2WorkspaceList(params: params))
-        case "workspace.create":
-            return v2Result(id: id, self.v2WorkspaceCreate(params: params))
-        case "workspace.select":
-            return v2Result(id: id, self.v2WorkspaceSelect(params: params))
-        case "workspace.current":
-            return v2Result(id: id, self.v2WorkspaceCurrent(params: params))
-        case "workspace.close":
-            return v2Result(id: id, self.v2WorkspaceClose(params: params))
-        case "workspace.move_to_window":
-            return v2Result(id: id, self.v2WorkspaceMoveToWindow(params: params))
-        case "workspace.reorder":
-            return v2Result(id: id, self.v2WorkspaceReorder(params: params))
-        case "workspace.rename":
-            return v2Result(id: id, self.v2WorkspaceRename(params: params))
-        case "workspace.action":
-            return v2Result(id: id, self.v2WorkspaceAction(params: params))
-        case "workspace.next":
-            return v2Result(id: id, self.v2WorkspaceNext(params: params))
-        case "workspace.previous":
-            return v2Result(id: id, self.v2WorkspacePrevious(params: params))
-        case "workspace.last":
-            return v2Result(id: id, self.v2WorkspaceLast(params: params))
-        case "workspace.remote.configure":
-            return v2Result(id: id, self.v2WorkspaceRemoteConfigure(params: params))
-        case "workspace.remote.reconnect":
-            return v2Result(id: id, self.v2WorkspaceRemoteReconnect(params: params))
-        case "workspace.remote.disconnect":
-            return v2Result(id: id, self.v2WorkspaceRemoteDisconnect(params: params))
-        case "workspace.remote.status":
-            return v2Result(id: id, self.v2WorkspaceRemoteStatus(params: params))
-        case "workspace.remote.terminal_session_end":
-            return v2Result(id: id, self.v2WorkspaceRemoteTerminalSessionEnd(params: params))
-        case "workspace.set_metadata":
-            return v2Result(id: id, self.v2WorkspaceSetMetadata(params: params))
-        case "workspace.get_metadata":
-            return v2Result(id: id, self.v2WorkspaceGetMetadata(params: params))
-        case "workspace.clear_metadata":
-            return v2Result(id: id, self.v2WorkspaceClearMetadata(params: params))
-        case "workspace.set_custom_color":
-            return v2Result(id: id, self.v2WorkspaceSetCustomColor(params: params))
-        case "workspace.apply":
-            return v2Result(id: id, self.v2WorkspaceApply(params: params))
 
         // Blueprints (CMUX-37 Phase 2)
-        case "workspace.list_blueprints":
-            return v2Result(id: id, self.v2WorkspaceListBlueprints(params: params))
-        case "workspace.export_blueprint":
-            return v2Result(id: id, self.v2WorkspaceExportBlueprint(params: params))
-        case "workspace.parse_blueprint":
-            return v2Result(id: id, self.v2WorkspaceParseBlueprint(params: params))
 
         // Snapshots (CMUX-37 Phase 1)
 
@@ -2932,32 +2882,6 @@ class TerminalController {
         // Mailbox
 
         // Panes
-        case "pane.list":
-            return v2Result(id: id, self.v2PaneList(params: params))
-        case "pane.focus":
-            return v2Result(id: id, self.v2PaneFocus(params: params))
-        case "pane.surfaces":
-            return v2Result(id: id, self.v2PaneSurfaces(params: params))
-        case "pane.create":
-            return v2Result(id: id, self.v2PaneCreate(params: params))
-        case "pane.resize":
-            return v2Result(id: id, self.v2PaneResize(params: params))
-        case "pane.swap":
-            return v2Result(id: id, self.v2PaneSwap(params: params))
-        case "pane.break":
-            return v2Result(id: id, self.v2PaneBreak(params: params))
-        case "pane.join":
-            return v2Result(id: id, self.v2PaneJoin(params: params))
-        case "pane.last":
-            return v2Result(id: id, self.v2PaneLast(params: params))
-        case "pane.confirm":
-            return v2Result(id: id, self.v2PaneConfirm(params: params))
-        case "pane.set_metadata":
-            return v2Result(id: id, self.v2PaneSetMetadata(params: params))
-        case "pane.get_metadata":
-            return v2Result(id: id, self.v2PaneGetMetadata(params: params))
-        case "pane.clear_metadata":
-            return v2Result(id: id, self.v2PaneClearMetadata(params: params))
 
         // Notifications
 
@@ -3239,6 +3163,8 @@ class TerminalController {
     /// unchanged. Runs on the main actor exactly like the switch it replaces.
     func v2DispatchExtracted(_ method: String, id: Any?, params: [String: Any]) -> String? {
         if method.hasPrefix("window.") { return v2DispatchWindow(method, id: id, params: params) }
+        if method.hasPrefix("workspace.") { return v2DispatchWorkspace(method, id: id, params: params) }
+        if method.hasPrefix("pane.") { return v2DispatchPane(method, id: id, params: params) }
         if method.hasPrefix("theme.") { return v2DispatchTheme(method, id: id, params: params) }
         if method.hasPrefix("system.") || method.hasPrefix("auth.") || method.hasPrefix("app.") { return v2DispatchSystem(method, id: id, params: params) }
         if method.hasPrefix("snapshot.") { return v2DispatchSnapshot(method, id: id, params: params) }
@@ -3943,767 +3869,14 @@ class TerminalController {
 
     // MARK: - V2 Workspace Methods
 
-    private func v2WorkspaceList(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
 
-        var workspaces: [[String: Any]] = []
-        v2MainSync {
-            workspaces = tabManager.tabs.enumerated().map { index, ws in
-                return [
-                    "id": ws.id.uuidString,
-                    "ref": v2Ref(kind: .workspace, uuid: ws.id),
-                    "index": index,
-                    "title": ws.title,
-                    "selected": ws.id == tabManager.selectedTabId,
-                    "pinned": ws.isPinned,
-                    "listening_ports": ws.listeningPorts,
-                    "remote": ws.remoteStatusPayload(),
-                    "current_directory": v2OrNull(ws.currentDirectory),
-                    "custom_color": v2OrNull(ws.customColor)
-                ]
-            }
-        }
 
-        let windowId = v2ResolveWindowId(tabManager: tabManager)
-        return .ok([
-            "window_id": v2OrNull(windowId?.uuidString),
-            "window_ref": v2Ref(kind: .window, uuid: windowId),
-            "workspaces": workspaces
-        ])
-    }
-    private func v2WorkspaceCreate(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
 
-        let requestedWorkingDirectory = v2RawString(params, "working_directory")?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let workingDirectory = (requestedWorkingDirectory?.isEmpty == false) ? requestedWorkingDirectory : nil
 
-        let requestedInitialCommand = v2RawString(params, "initial_command")?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let initialCommand = (requestedInitialCommand?.isEmpty == false) ? requestedInitialCommand : nil
 
-        let rawInitialEnv = v2StringMap(params, "initial_env") ?? [:]
-        let initialEnv = rawInitialEnv.reduce(into: [String: String]()) { result, pair in
-            let key = pair.key.trimmingCharacters(in: .whitespacesAndNewlines)
-            guard !key.isEmpty else { return }
-            result[key] = pair.value
-        }
-        let cwd: String?
-        if let workingDirectory {
-            cwd = workingDirectory
-        } else if let raw = params["cwd"] {
-            guard let str = raw as? String else {
-                return .err(code: "invalid_params", message: "cwd must be a string", data: nil)
-            }
-            cwd = str
-        } else {
-            cwd = nil
-        }
 
-        // Layout path: let WorkspaceLayoutExecutor own creation so exactly one workspace is created.
-        // Pre-creating a workspace here causes a ghost tab on every successful layout apply (B1).
-        if let layoutDict = params["layout"] as? [String: Any], !layoutDict.isEmpty {
-            let layoutResult = v2WorkspaceApply(params: layoutDict)
 
-            if case .err = layoutResult {
-                return layoutResult
-            }
 
-            guard case .ok(let resultAny) = layoutResult,
-                  let resultDict = resultAny as? [String: Any] else {
-                return .err(code: "internal_error", message: "Unexpected apply result", data: nil)
-            }
-
-            // Transactional: non-empty failures mean partial apply — roll back the workspace.
-            let failures = resultDict["failures"] as? [[String: Any]] ?? []
-            if !failures.isEmpty {
-                if let refStr = resultDict["workspaceRef"] as? String,
-                   let wsUUID = v2ResolveHandleRef(refStr) {
-                    v2MainSync {
-                        if let ws = tabManager.tabs.first(where: { $0.id == wsUUID }) {
-                            tabManager.closeWorkspace(ws)
-                        }
-                    }
-                }
-                return .err(code: "apply_failed", message: "Layout application partially failed", data: resultDict)
-            }
-
-            // Normalize to the standard snake_case workspace creation envelope (B2).
-            let workspaceRef = resultDict["workspaceRef"] as? String ?? ""
-            let wsUUID = v2ResolveHandleRef(workspaceRef)
-
-            // An explicit --title wins over any title the blueprint set, and is
-            // applied via the same customTitle path that workspace.rename uses.
-            let layoutTitle = v2RawString(params, "title")?.trimmingCharacters(in: .whitespacesAndNewlines)
-            let appliedLayoutTitle = (layoutTitle?.isEmpty == false) ? layoutTitle : nil
-            if let appliedLayoutTitle, let wsUUID {
-                v2MainSync {
-                    if tabManager.tabs.contains(where: { $0.id == wsUUID }) {
-                        tabManager.setCustomTitle(tabId: wsUUID, title: appliedLayoutTitle)
-                    }
-                }
-            }
-
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            return .ok([
-                "workspace_id": v2OrNull(wsUUID?.uuidString),
-                "workspace_ref": workspaceRef as Any,
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "title": v2OrNull(appliedLayoutTitle),
-                "layout_result": resultAny
-            ])
-        }
-
-        // Optional title: set the same customTitle field that workspace.rename writes,
-        // so `new-workspace --title` is atomic with creation (no follow-up rename).
-        let requestedTitle = v2RawString(params, "title")?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let customTitle = (requestedTitle?.isEmpty == false) ? requestedTitle : nil
-
-        // No layout: create workspace normally.
-        var newId: UUID?
-        let shouldFocus = v2FocusAllowed()
-        guard v2MainSyncWithDeadline({
-            let ws = tabManager.addWorkspace(
-                workingDirectory: cwd,
-                initialTerminalCommand: initialCommand,
-                initialTerminalEnvironment: initialEnv,
-                select: shouldFocus,
-                eagerLoadTerminal: !shouldFocus
-            )
-            newId = ws.id
-            if let customTitle {
-                tabManager.setCustomTitle(tabId: ws.id, title: customTitle)
-            }
-            return
-        }) != nil else {
-            return .err(code: "main_thread_timeout", message: "main thread did not respond within deadline", data: nil)
-        }
-
-        guard let newId else {
-            return .err(code: "internal_error", message: "Failed to create workspace", data: nil)
-        }
-
-        let windowId = v2ResolveWindowId(tabManager: tabManager)
-        return .ok([
-            "window_id": v2OrNull(windowId?.uuidString),
-            "window_ref": v2Ref(kind: .window, uuid: windowId),
-            "workspace_id": newId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: newId),
-            "title": v2OrNull(customTitle)
-        ])
-    }
-    private func v2WorkspaceSelect(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let wsId = v2UUID(params, "workspace_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-
-        var success = false
-        v2MainSync {
-            if let ws = tabManager.tabs.first(where: { $0.id == wsId }) {
-                // If this workspace belongs to another window, bring it forward so focus is visible.
-                if let windowId = v2ResolveWindowId(tabManager: tabManager) {
-                    _ = AppDelegate.shared?.focusMainWindow(windowId: windowId)
-                    setActiveTabManager(tabManager)
-                    // Bring c11 to the macOS foreground for explicit focus-intent commands.
-                    // workspace.select is in focusIntentV2Methods, so this is intentional.
-                    DispatchQueue.main.async {
-                        NSApp.activate(ignoringOtherApps: true)
-                    }
-                }
-                tabManager.selectWorkspace(ws)
-                success = true
-            }
-        }
-
-        let windowId = v2ResolveWindowId(tabManager: tabManager)
-        return success
-            ? .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": wsId.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: wsId)
-            ])
-            : .err(code: "not_found", message: "Workspace not found", data: [
-                "workspace_id": wsId.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: wsId)
-            ])
-    }
-    private func v2WorkspaceCurrent(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        var wsId: UUID?
-        var wsPayload: [String: Any]?
-        v2MainSync {
-            wsId = tabManager.selectedTabId
-            if let wsId, let workspace = tabManager.tabs.first(where: { $0.id == wsId }) {
-                wsPayload = [
-                    "id": workspace.id.uuidString,
-                    "ref": v2Ref(kind: .workspace, uuid: workspace.id),
-                    "title": workspace.title,
-                    "selected": true,
-                    "pinned": workspace.isPinned,
-                    "listening_ports": workspace.listeningPorts,
-                    "remote": workspace.remoteStatusPayload(),
-                ]
-            }
-        }
-        guard let wsId else {
-            return .err(code: "not_found", message: "No workspace selected", data: nil)
-        }
-        let windowId = v2ResolveWindowId(tabManager: tabManager)
-        return .ok([
-            "window_id": v2OrNull(windowId?.uuidString),
-            "window_ref": v2Ref(kind: .window, uuid: windowId),
-            "workspace_id": wsId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: wsId),
-            "workspace": wsPayload ?? NSNull()
-        ])
-    }
-    private func v2WorkspaceClose(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let wsId = v2UUID(params, "workspace_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-
-        var found = false
-        v2MainSync {
-            if let ws = tabManager.tabs.first(where: { $0.id == wsId }) {
-                tabManager.closeWorkspace(ws)
-                found = true
-            }
-        }
-
-        let windowId = v2ResolveWindowId(tabManager: tabManager)
-        return found
-            ? .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": wsId.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: wsId)
-            ])
-            : .err(code: "not_found", message: "Workspace not found", data: [
-                "workspace_id": wsId.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: wsId)
-            ])
-    }
-    private func v2WorkspaceMoveToWindow(params: [String: Any]) -> V2CallResult {
-        guard let wsId = v2UUID(params, "workspace_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-        guard let windowId = v2UUID(params, "window_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid window_id", data: nil)
-        }
-        let focus = v2FocusAllowed(requested: v2Bool(params, "focus") ?? false)
-
-        var result: V2CallResult = .err(code: "internal_error", message: "Failed to move workspace", data: nil)
-        v2MainSync {
-            guard let srcTM = AppDelegate.shared?.tabManagerFor(tabId: wsId) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: ["workspace_id": wsId.uuidString])
-                return
-            }
-            guard let dstTM = AppDelegate.shared?.tabManagerFor(windowId: windowId) else {
-                result = .err(code: "not_found", message: "Window not found", data: ["window_id": windowId.uuidString])
-                return
-            }
-            guard let ws = srcTM.detachWorkspace(tabId: wsId) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: ["workspace_id": wsId.uuidString])
-                return
-            }
-
-            dstTM.attachWorkspace(ws, select: focus)
-            if focus {
-                _ = AppDelegate.shared?.focusMainWindow(windowId: windowId)
-                setActiveTabManager(dstTM)
-            }
-            result = .ok([
-                "workspace_id": wsId.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: wsId),
-                "window_id": windowId.uuidString,
-                "window_ref": v2Ref(kind: .window, uuid: windowId)
-            ])
-        }
-        return result
-    }
-    private func v2WorkspaceReorder(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let workspaceId = v2UUID(params, "workspace_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-
-        let index = v2Int(params, "index")
-        let beforeId = v2UUID(params, "before_workspace_id")
-        let afterId = v2UUID(params, "after_workspace_id")
-
-        let targetCount = (index != nil ? 1 : 0) + (beforeId != nil ? 1 : 0) + (afterId != nil ? 1 : 0)
-        if targetCount != 1 {
-            return .err(
-                code: "invalid_params",
-                message: "Specify exactly one target: index, before_workspace_id, or after_workspace_id",
-                data: nil
-            )
-        }
-
-        var moved = false
-        var newIndex: Int?
-        v2MainSync {
-            if let index {
-                moved = tabManager.reorderWorkspace(tabId: workspaceId, toIndex: index)
-            } else {
-                moved = tabManager.reorderWorkspace(tabId: workspaceId, before: beforeId, after: afterId)
-            }
-            newIndex = tabManager.tabs.firstIndex(where: { $0.id == workspaceId })
-        }
-
-        guard moved else {
-            return .err(code: "not_found", message: "Workspace not found", data: ["workspace_id": workspaceId.uuidString])
-        }
-
-        let windowId = v2ResolveWindowId(tabManager: tabManager)
-        return .ok([
-            "workspace_id": workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-            "window_id": v2OrNull(windowId?.uuidString),
-            "window_ref": v2Ref(kind: .window, uuid: windowId),
-            "index": v2OrNull(newIndex)
-        ])
-    }
-    private func v2WorkspaceSetCustomColor(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let workspaceId = v2UUID(params, "workspace_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-
-        // Accept either a hex string to set, or `clear: true` to reset.
-        let clear = (params["clear"] as? Bool) ?? false
-        let hex = params["hex"] as? String
-
-        if !clear && hex == nil {
-            return .err(code: "invalid_params", message: "Provide either 'hex' or 'clear=true'", data: nil)
-        }
-
-        var applied: String? = nil
-        var found = false
-        v2MainSync {
-            guard let workspace = tabManager.tabs.first(where: { $0.id == workspaceId }) else { return }
-            found = true
-            if clear {
-                workspace.setCustomColor(nil)
-                applied = nil
-            } else if let hex {
-                workspace.setCustomColor(hex)
-                applied = workspace.customColor
-            }
-        }
-
-        guard found else {
-            return .err(code: "not_found", message: "Workspace not found", data: [
-                "workspace_id": workspaceId.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId)
-            ])
-        }
-
-        return .ok([
-            "workspace_id": workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-            "hex": v2OrNull(applied),
-            "cleared": clear
-        ])
-    }
-
-    private func v2WorkspaceRename(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let workspaceId = v2UUID(params, "workspace_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-        guard let titleRaw = v2String(params, "title"),
-              !titleRaw.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty else {
-            return .err(code: "invalid_params", message: "Missing or invalid title", data: nil)
-        }
-
-        let title = titleRaw.trimmingCharacters(in: .whitespacesAndNewlines)
-        var renamed = false
-        v2MainSync {
-            guard tabManager.tabs.contains(where: { $0.id == workspaceId }) else { return }
-            tabManager.setCustomTitle(tabId: workspaceId, title: title)
-            renamed = true
-        }
-
-        guard renamed else {
-            return .err(code: "not_found", message: "Workspace not found", data: [
-                "workspace_id": workspaceId.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId)
-            ])
-        }
-
-        let windowId = v2ResolveWindowId(tabManager: tabManager)
-        return .ok([
-            "workspace_id": workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-            "window_id": v2OrNull(windowId?.uuidString),
-            "window_ref": v2Ref(kind: .window, uuid: windowId),
-            "title": title
-        ])
-    }
-    private func v2WorkspaceNext(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-
-        var result: V2CallResult = .err(code: "not_found", message: "No workspace selected", data: nil)
-        v2MainSync {
-            guard tabManager.selectedTabId != nil else { return }
-            if let windowId = v2ResolveWindowId(tabManager: tabManager) {
-                _ = AppDelegate.shared?.focusMainWindow(windowId: windowId)
-                setActiveTabManager(tabManager)
-            }
-            tabManager.selectNextTab()
-            guard let workspaceId = tabManager.selectedTabId else { return }
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            result = .ok([
-                "workspace_id": workspaceId.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId)
-            ])
-        }
-        return result
-    }
-
-    private func v2WorkspacePrevious(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-
-        var result: V2CallResult = .err(code: "not_found", message: "No workspace selected", data: nil)
-        v2MainSync {
-            guard tabManager.selectedTabId != nil else { return }
-            if let windowId = v2ResolveWindowId(tabManager: tabManager) {
-                _ = AppDelegate.shared?.focusMainWindow(windowId: windowId)
-                setActiveTabManager(tabManager)
-            }
-            tabManager.selectPreviousTab()
-            guard let workspaceId = tabManager.selectedTabId else { return }
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            result = .ok([
-                "workspace_id": workspaceId.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId)
-            ])
-        }
-        return result
-    }
-
-    private func v2WorkspaceLast(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-
-        var result: V2CallResult = .err(code: "not_found", message: "No previous workspace in history", data: nil)
-        v2MainSync {
-            guard let before = tabManager.selectedTabId else { return }
-            if let windowId = v2ResolveWindowId(tabManager: tabManager) {
-                _ = AppDelegate.shared?.focusMainWindow(windowId: windowId)
-                setActiveTabManager(tabManager)
-            }
-            tabManager.navigateBack()
-            guard let after = tabManager.selectedTabId, after != before else { return }
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            result = .ok([
-                "workspace_id": after.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: after),
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId)
-            ])
-        }
-        return result
-    }
-
-    private func v2WorkspaceRemoteConfigure(params: [String: Any]) -> V2CallResult {
-        let requestedWorkspaceId = v2UUID(params, "workspace_id")
-        if v2HasNonNullParam(params, "workspace_id"), requestedWorkspaceId == nil {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-        let fallbackTabManager = v2ResolveTabManager(params: params)
-        let workspaceId = requestedWorkspaceId ?? fallbackTabManager?.selectedTabId
-        guard let workspaceId else {
-            return .err(code: "invalid_params", message: "Missing workspace_id", data: nil)
-        }
-        guard let destination = v2String(params, "destination") else {
-            return .err(code: "invalid_params", message: "Missing destination", data: nil)
-        }
-
-        var sshPort: Int?
-        if v2HasNonNullParam(params, "port") {
-            guard let parsedPort = v2StrictInt(params, "port"),
-                  parsedPort > 0,
-                  parsedPort <= 65535 else {
-                return .err(code: "invalid_params", message: "port must be 1-65535", data: nil)
-            }
-            sshPort = parsedPort
-        }
-
-        // Internal deterministic test hook: pin the local proxy listener port to force bind conflicts.
-        var localProxyPort: Int?
-        if v2HasNonNullParam(params, "local_proxy_port") {
-            guard let parsedLocalProxyPort = v2StrictInt(params, "local_proxy_port"),
-                  parsedLocalProxyPort > 0,
-                  parsedLocalProxyPort <= 65535 else {
-                return .err(code: "invalid_params", message: "local_proxy_port must be 1-65535", data: nil)
-            }
-            localProxyPort = parsedLocalProxyPort
-        }
-
-        let identityFile = v2RawString(params, "identity_file")?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let sshOptions = v2StringArray(params, "ssh_options") ?? []
-        let autoConnect = v2Bool(params, "auto_connect") ?? true
-        var relayPort: Int?
-        if v2HasNonNullParam(params, "relay_port") {
-            guard let parsedRelayPort = v2StrictInt(params, "relay_port"),
-                  parsedRelayPort > 0,
-                  parsedRelayPort <= 65535 else {
-                return .err(code: "invalid_params", message: "relay_port must be 1-65535", data: nil)
-            }
-            relayPort = parsedRelayPort
-        }
-        let relayID = v2RawString(params, "relay_id")?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let relayToken = v2RawString(params, "relay_token")?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let localSocketPath = v2RawString(params, "local_socket_path")
-        let terminalStartupCommand = v2RawString(params, "terminal_startup_command")?
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-        if relayPort != nil {
-            guard let relayID, !relayID.isEmpty else {
-                return .err(code: "invalid_params", message: "relay_id is required when relay_port is set", data: nil)
-            }
-            guard let relayToken,
-                  relayToken.range(of: "^[0-9a-f]{64}$", options: .regularExpression) != nil else {
-                return .err(code: "invalid_params", message: "relay_token must be 64 lowercase hex characters when relay_port is set", data: nil)
-            }
-        }
-
-#if DEBUG
-        dlog(
-            "workspace.remote.configure.request workspace=\(workspaceId.uuidString.prefix(8)) " +
-            "target=\(destination) port=\(sshPort.map(String.init) ?? "nil") " +
-            "autoConnect=\(autoConnect ? 1 : 0) relayPort=\(relayPort.map(String.init) ?? "nil") " +
-            "localSocket=\(localSocketPath?.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty == false ? localSocketPath! : "nil") " +
-            "sshOptions=\(sshOptions.joined(separator: "|"))"
-        )
-#endif
-        var result: V2CallResult = .err(code: "not_found", message: "Workspace not found", data: [
-            "workspace_id": workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-        ])
-
-        // Must run on main for v2MainSync because Workspace.configureRemoteConnection mutates TabManager/UI-owned workspace state.
-        v2MainSync {
-            guard let owner = AppDelegate.shared?.tabManagerFor(tabId: workspaceId),
-                  let workspace = owner.tabs.first(where: { $0.id == workspaceId }) else {
-                return
-            }
-
-            let config = WorkspaceRemoteConfiguration(
-                destination: destination,
-                port: sshPort,
-                identityFile: identityFile?.isEmpty == true ? nil : identityFile,
-                sshOptions: sshOptions,
-                localProxyPort: localProxyPort,
-                relayPort: relayPort,
-                relayID: relayID?.isEmpty == true ? nil : relayID,
-                relayToken: relayToken?.isEmpty == true ? nil : relayToken,
-                localSocketPath: localSocketPath,
-                terminalStartupCommand: terminalStartupCommand?.isEmpty == true ? nil : terminalStartupCommand
-            )
-            workspace.configureRemoteConnection(config, autoConnect: autoConnect)
-
-            let windowId = v2ResolveWindowId(tabManager: owner)
-            result = .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": workspace.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: workspace.id),
-                "remote": workspace.remoteStatusPayload(),
-            ])
-        }
-
-        return result
-    }
-
-    private func v2WorkspaceRemoteDisconnect(params: [String: Any]) -> V2CallResult {
-        let requestedWorkspaceId = v2UUID(params, "workspace_id")
-        if v2HasNonNullParam(params, "workspace_id"), requestedWorkspaceId == nil {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-        let fallbackTabManager = v2ResolveTabManager(params: params)
-        let workspaceId = requestedWorkspaceId ?? fallbackTabManager?.selectedTabId
-        guard let workspaceId else {
-            return .err(code: "invalid_params", message: "Missing workspace_id", data: nil)
-        }
-
-        let clearConfiguration = v2Bool(params, "clear") ?? false
-        var result: V2CallResult = .err(code: "not_found", message: "Workspace not found", data: [
-            "workspace_id": workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-        ])
-
-        // Must run on main for v2MainSync because disconnect mutates TabManager/UI-owned workspace state.
-        v2MainSync {
-            guard let owner = AppDelegate.shared?.tabManagerFor(tabId: workspaceId),
-                  let workspace = owner.tabs.first(where: { $0.id == workspaceId }) else {
-                return
-            }
-
-            workspace.disconnectRemoteConnection(clearConfiguration: clearConfiguration)
-            let windowId = v2ResolveWindowId(tabManager: owner)
-            result = .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": workspace.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: workspace.id),
-                "remote": workspace.remoteStatusPayload(),
-            ])
-        }
-
-        return result
-    }
-
-    private func v2WorkspaceRemoteReconnect(params: [String: Any]) -> V2CallResult {
-        let requestedWorkspaceId = v2UUID(params, "workspace_id")
-        if v2HasNonNullParam(params, "workspace_id"), requestedWorkspaceId == nil {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-        let fallbackTabManager = v2ResolveTabManager(params: params)
-        let workspaceId = requestedWorkspaceId ?? fallbackTabManager?.selectedTabId
-        guard let workspaceId else {
-            return .err(code: "invalid_params", message: "Missing workspace_id", data: nil)
-        }
-
-        var result: V2CallResult = .err(code: "not_found", message: "Workspace not found", data: [
-            "workspace_id": workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-        ])
-
-        // Must run on main for v2MainSync because reconnect mutates TabManager/UI-owned workspace state.
-        v2MainSync {
-            guard let owner = AppDelegate.shared?.tabManagerFor(tabId: workspaceId),
-                  let workspace = owner.tabs.first(where: { $0.id == workspaceId }) else {
-                return
-            }
-
-            guard workspace.remoteConfiguration != nil else {
-                result = .err(code: "invalid_state", message: "Remote workspace is not configured", data: [
-                    "workspace_id": workspaceId.uuidString,
-                    "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-                ])
-                return
-            }
-
-            workspace.reconnectRemoteConnection()
-            let windowId = v2ResolveWindowId(tabManager: owner)
-            result = .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": workspace.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: workspace.id),
-                "remote": workspace.remoteStatusPayload(),
-            ])
-        }
-
-        return result
-    }
-
-    private func v2WorkspaceRemoteStatus(params: [String: Any]) -> V2CallResult {
-        let requestedWorkspaceId = v2UUID(params, "workspace_id")
-        if v2HasNonNullParam(params, "workspace_id"), requestedWorkspaceId == nil {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-        let fallbackTabManager = v2ResolveTabManager(params: params)
-        let workspaceId = requestedWorkspaceId ?? fallbackTabManager?.selectedTabId
-        guard let workspaceId else {
-            return .err(code: "invalid_params", message: "Missing workspace_id", data: nil)
-        }
-
-        var result: V2CallResult = .err(code: "not_found", message: "Workspace not found", data: [
-            "workspace_id": workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-        ])
-
-        // Must run on main for v2MainSync because Workspace.remoteStatusPayload reads TabManager/UI-owned state.
-        v2MainSync {
-            guard let owner = AppDelegate.shared?.tabManagerFor(tabId: workspaceId),
-                  let workspace = owner.tabs.first(where: { $0.id == workspaceId }) else {
-                return
-            }
-            let windowId = v2ResolveWindowId(tabManager: owner)
-            result = .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": workspace.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: workspace.id),
-                "remote": workspace.remoteStatusPayload(),
-            ])
-        }
-
-        return result
-    }
-
-    private func v2WorkspaceRemoteTerminalSessionEnd(params: [String: Any]) -> V2CallResult {
-        guard let workspaceId = v2UUID(params, "workspace_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid workspace_id", data: nil)
-        }
-        guard let surfaceId = v2UUID(params, "surface_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid surface_id", data: nil)
-        }
-        guard let relayPort = v2StrictInt(params, "relay_port"),
-              relayPort > 0,
-              relayPort <= 65535 else {
-            return .err(code: "invalid_params", message: "Missing or invalid relay_port", data: nil)
-        }
-
-        var result: V2CallResult = .err(code: "not_found", message: "Workspace not found", data: [
-            "workspace_id": workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: workspaceId),
-            "surface_id": surfaceId.uuidString,
-            "surface_ref": v2Ref(kind: .surface, uuid: surfaceId),
-            "relay_port": relayPort,
-        ])
-
-        v2MainSync {
-            guard let owner = AppDelegate.shared?.tabManagerFor(tabId: workspaceId),
-                  let workspace = owner.tabs.first(where: { $0.id == workspaceId }) else {
-                return
-            }
-            workspace.markRemoteTerminalSessionEnded(surfaceId: surfaceId, relayPort: relayPort)
-            let windowId = v2ResolveWindowId(tabManager: owner)
-            result = .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": workspace.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: workspace.id),
-                "surface_id": surfaceId.uuidString,
-                "surface_ref": v2Ref(kind: .surface, uuid: surfaceId),
-                "relay_port": relayPort,
-                "remote": workspace.remoteStatusPayload(),
-            ])
-        }
-
-        return result
-    }
 
     /// Resolve the Workspace referenced by `workspace_id`. Falls back to the
     /// currently selected workspace on the caller's TabManager if no explicit
@@ -4731,270 +3904,11 @@ class TerminalController {
         }
     }
 
-    /// CMUX-37 Phase 0: `workspace.apply` v2 handler. Decodes a
-    /// `WorkspaceApplyPlan` from params, runs it through
-    /// `WorkspaceLayoutExecutor` on the main actor, and returns the
-    /// `ApplyResult` as a JSON object. The debug/test surface for the
-    /// Phase 0 primitive — Blueprints (Phase 2) and Snapshots (Phase 1)
-    /// layer on top of the same executor.
-    private func v2WorkspaceApply(params: [String: Any]) -> V2CallResult {
-        // Decode the plan and (optional) options off-main. Validation
-        // failures never touch the main actor.
-        guard let planRaw = params["plan"] else {
-            return .err(code: "invalid_params", message: "Missing 'plan'", data: nil)
-        }
-        let plan: WorkspaceApplyPlan
-        do {
-            let planData = try JSONSerialization.data(withJSONObject: planRaw, options: [])
-            plan = try JSONDecoder().decode(WorkspaceApplyPlan.self, from: planData)
-        } catch {
-            return .err(
-                code: "invalid_params",
-                message: "Failed to decode WorkspaceApplyPlan: \(error)",
-                data: nil
-            )
-        }
-
-        let options: ApplyOptions
-        if let optionsRaw = params["options"] {
-            do {
-                let optionsData = try JSONSerialization.data(withJSONObject: optionsRaw, options: [])
-                options = try JSONDecoder().decode(ApplyOptions.self, from: optionsData)
-            } catch {
-                return .err(
-                    code: "invalid_params",
-                    message: "Failed to decode ApplyOptions: \(error)",
-                    data: nil
-                )
-            }
-        } else {
-            options = ApplyOptions()
-        }
-
-        // Pre-validate off-main. Per review cycle 1 I3, the v2 handler
-        // must not ride MainActor for pure validation — this block runs
-        // entirely on the socket worker thread. A malformed plan returns
-        // a typed `invalid_params` error (JSON-RPC clients parse as
-        // failure), not an `ok` envelope with an ApplyFailure body —
-        // cycle 2 IM2 fix.
-        if let validationFailure = WorkspaceLayoutExecutor.validate(plan: plan) {
-            let data: [String: Any] = [
-                "failure": [
-                    "code": validationFailure.code,
-                    "step": validationFailure.step,
-                    "message": validationFailure.message
-                ]
-            ]
-            return .err(
-                code: "invalid_params",
-                message: "WorkspaceApplyPlan validation failed: \(validationFailure.message)",
-                data: data
-            )
-        }
-
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-
-        // Apply the socket focus policy (CLAUDE.md: "Socket/CLI commands
-        // must not steal macOS app focus"). workspace.apply is not in
-        // `focusIntentV2Methods`, so `v2FocusAllowed` returns false for
-        // this command regardless of caller intent. Zero out
-        // `options.select` before dispatch so the executor does not
-        // raise the window / select the created workspace. Cycle 2 IM1
-        // fix.
-        var effectiveOptions = options
-        if options.select && !v2FocusAllowed(requested: true) {
-            effectiveOptions.select = false
-        }
-
-        var result: ApplyResult?
-        v2MainSync {
-            let deps = WorkspaceLayoutExecutorDependencies(
-                tabManager: tabManager,
-                workspaceRefMinter: { [weak self] uuid in
-                    self?.v2EnsureHandleRef(kind: .workspace, uuid: uuid) ?? "workspace:\(uuid.uuidString)"
-                },
-                surfaceRefMinter: { [weak self] uuid in
-                    self?.v2EnsureHandleRef(kind: .surface, uuid: uuid) ?? "surface:\(uuid.uuidString)"
-                },
-                paneRefMinter: { [weak self] uuid in
-                    self?.v2EnsureHandleRef(kind: .pane, uuid: uuid) ?? "pane:\(uuid.uuidString)"
-                }
-            )
-            result = WorkspaceLayoutExecutor.apply(plan, options: effectiveOptions, dependencies: deps)
-        }
-
-        guard let applyResult = result else {
-            return .err(code: "internal_error", message: "Executor returned no result", data: nil)
-        }
-
-        // Encode ApplyResult back to [String: Any] for the v2 JSON envelope.
-        do {
-            let encoded = try JSONEncoder().encode(applyResult)
-            let asAny = try JSONSerialization.jsonObject(with: encoded, options: [])
-            return .ok(asAny)
-        } catch {
-            return .err(
-                code: "internal_error",
-                message: "Failed to encode ApplyResult: \(error)",
-                data: nil
-            )
-        }
-    }
 
     // MARK: - V2 Blueprint Methods (CMUX-37 Phase 2)
 
-    /// `workspace.list_blueprints`: returns a merged index of all discovered
-    /// blueprint files. Optional `cwd` param enables per-repo discovery.
-    /// Returns `{blueprints: [...WorkspaceBlueprintIndex...]}`.
-    private func v2WorkspaceListBlueprints(params: [String: Any]) -> V2CallResult {
-        let cwdURL: URL?
-        if let cwdStr = params["cwd"] as? String, !cwdStr.isEmpty {
-            cwdURL = URL(fileURLWithPath: cwdStr)
-        } else {
-            cwdURL = nil
-        }
-        let store = WorkspaceBlueprintStore()
-        let entries = store.merged(cwd: cwdURL)
-        let encoder = JSONEncoder()
-        encoder.dateEncodingStrategy = .iso8601
-        do {
-            let encoded = try encoder.encode(entries)
-            let asAny = try JSONSerialization.jsonObject(with: encoded, options: [])
-            return .ok(["blueprints": asAny])
-        } catch {
-            return .err(code: "internal_error", message: "\(error)", data: nil)
-        }
-    }
 
-    /// `workspace.export_blueprint`: captures the live workspace as a
-    /// `WorkspaceBlueprintFile` and writes it to the user blueprint directory
-    /// under `~/.config/c11/blueprints/<sanitized-name>.<ext>` (default
-    /// extension `.md`; `.json` selectable via the `format` param).
-    /// Required param: `name`. Optional: `workspace_id`, `description`,
-    /// `force` (bool), `format` (`"md"` default | `"json"`).
-    /// Returns `{path: "...", name: "...", format: "md|json"}`.
-    private func v2WorkspaceExportBlueprint(params: [String: Any]) -> V2CallResult {
-        guard let name = params["name"] as? String, !name.isEmpty else {
-            return .err(code: "invalid_params", message: "Missing or empty 'name'", data: nil)
-        }
-        let description = params["description"] as? String
-        let force = params["force"] as? Bool ?? false
-        let formatRaw = (params["format"] as? String)?.lowercased() ?? "md"
-        let formatExt: String
-        switch formatRaw {
-        case "md", "markdown": formatExt = "md"
-        case "json":           formatExt = "json"
-        default:
-            return .err(
-                code: "invalid_params",
-                message: "Unsupported format '\(formatRaw)'. Expected 'md' or 'json'.",
-                data: nil
-            )
-        }
 
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-
-        var blueprintFile: WorkspaceBlueprintFile?
-        v2MainSync {
-            let exporter = WorkspaceBlueprintExporter(tabManager: tabManager)
-            guard let workspace = v2ResolveWorkspace(params: params, tabManager: tabManager) else { return }
-            blueprintFile = exporter.export(workspaceId: workspace.id, name: name, description: description)
-        }
-
-        guard let file = blueprintFile else {
-            return .err(code: "not_found", message: "Workspace not found", data: nil)
-        }
-
-        // Sanitize name to a filesystem-safe filename.
-        let sanitized = String(name
-            .replacingOccurrences(of: " ", with: "-")
-            .filter { $0.isLetter || $0.isNumber || $0 == "-" || $0 == "_" || $0 == "." })
-        let filename = sanitized.isEmpty ? "blueprint" : sanitized
-
-        let fm = FileManager.default
-        let dir = fm.homeDirectoryForCurrentUser
-            .appendingPathComponent(".config/c11/blueprints")
-        let url = dir.appendingPathComponent("\(filename).\(formatExt)")
-
-        if !force && fm.fileExists(atPath: url.path) {
-            return .err(
-                code: "BLUEPRINT_ALREADY_EXISTS",
-                message: "A blueprint named '\(name)' already exists at \(url.path). Use --force to overwrite.",
-                data: nil
-            )
-        }
-
-        let store = WorkspaceBlueprintStore()
-        do {
-            try store.write(file, to: url)
-        } catch let err as WorkspaceBlueprintStore.StoreError {
-            return .err(code: err.code, message: "\(err)", data: nil)
-        } catch {
-            return .err(code: "export_failed", message: "\(error)", data: nil)
-        }
-
-        return .ok(["path": url.path, "name": name, "format": formatExt])
-    }
-
-    /// `workspace.parse_blueprint`: pure parser for a markdown or JSON
-    /// blueprint envelope passed by content. Required params: `content`
-    /// (string), `format` (`"md"` or `"json"`). Returns the decoded
-    /// envelope as `{name, description, plan}` so the CLI can forward
-    /// `plan` to `workspace.apply` without bringing the parser into the
-    /// CLI binary's link surface.
-    ///
-    /// File reads happen in the CLI process (caller's real FS
-    /// permissions); the socket only sees content already in memory. No
-    /// disk access here, no path resolution — keeps this immune to the
-    /// arbitrary-file-read class of attacks `snapshot.restore` already
-    /// guards against.
-    private func v2WorkspaceParseBlueprint(params: [String: Any]) -> V2CallResult {
-        guard let content = params["content"] as? String else {
-            return .err(code: "invalid_params", message: "Missing 'content'", data: nil)
-        }
-        let formatRaw = ((params["format"] as? String) ?? "md").lowercased()
-        guard let data = content.data(using: .utf8) else {
-            return .err(code: "invalid_params", message: "Content is not valid UTF-8", data: nil)
-        }
-        let file: WorkspaceBlueprintFile
-        do {
-            switch formatRaw {
-            case "md", "markdown":
-                file = try WorkspaceBlueprintMarkdown.parse(data)
-            case "json":
-                file = try JSONDecoder().decode(WorkspaceBlueprintFile.self, from: data)
-            default:
-                return .err(
-                    code: "invalid_params",
-                    message: "Unsupported format '\(formatRaw)'. Expected 'md' or 'json'.",
-                    data: nil
-                )
-            }
-        } catch {
-            return .err(code: "blueprint_decode_failed", message: "\(error)", data: nil)
-        }
-        do {
-            let encoded = try JSONEncoder().encode(file)
-            guard let asAny = try JSONSerialization.jsonObject(with: encoded, options: []) as? [String: Any],
-                  let plan = asAny["plan"] else {
-                return .err(code: "internal_error", message: "Failed to materialize parsed blueprint", data: nil)
-            }
-            var payload: [String: Any] = [
-                "plan": plan,
-                "name": file.name
-            ]
-            if let desc = file.description {
-                payload["description"] = desc
-            }
-            return .ok(payload)
-        } catch {
-            return .err(code: "internal_error", message: "Failed to encode parsed blueprint: \(error)", data: nil)
-        }
-    }
 
     // MARK: - V2 Snapshot Methods (CMUX-37 Phase 1)
 
@@ -5003,344 +3917,9 @@ class TerminalController {
 
 
 
-    private func v2WorkspaceSetMetadata(params: [String: Any]) -> V2CallResult {
-        // Parse + validate off-main per the socket command threading policy
-        // (CLAUDE.md "Socket command threading policy").
-        let rawMetadata = v2StringMap(params, "metadata")
-        let singleKey = v2String(params, "key")
-        let singleValueRaw = params["value"]
 
-        var writes: [String: String?] = [:]
-        if let rawMetadata {
-            for (k, v) in rawMetadata { writes[k] = v }
-        }
-        if let singleKey {
-            if singleValueRaw is NSNull {
-                writes[singleKey] = nil
-            } else if let s = singleValueRaw as? String {
-                writes[singleKey] = s
-            } else if singleValueRaw == nil {
-                return .err(code: "invalid_params", message: "Missing 'value' for key '\(singleKey)'", data: nil)
-            } else {
-                return .err(code: "invalid_params", message: "metadata value must be a string or null", data: nil)
-            }
-        }
 
-        if writes.isEmpty {
-            return .err(
-                code: "invalid_params",
-                message: "Provide 'metadata' object or 'key'/'value' pair",
-                data: nil
-            )
-        }
 
-        // Validate all keys and non-nil values off-main before taking the
-        // main-actor critical section.
-        for (k, maybeValue) in writes {
-            do {
-                if let value = maybeValue {
-                    try WorkspaceMetadataValidator.validate(key: k, value: value)
-                } else {
-                    try WorkspaceMetadataValidator.validateKey(k)
-                }
-            } catch let err as WorkspaceMetadataValidator.ValidationError {
-                return .err(code: err.code, message: err.message, data: err.detail)
-            } catch {
-                return .err(code: "internal_error", message: "\(error)", data: nil)
-            }
-        }
-
-        guard let resolved = v2ResolveWorkspaceForMetadata(params: params) else {
-            return .err(code: "not_found", message: "Workspace not found", data: nil)
-        }
-
-        var capacityError: WorkspaceMetadataValidator.CapacityError?
-        var resultMetadata: [String: String] = [:]
-        // Workspace is @MainActor; the mutation must run on the main actor.
-        // Precedent: workspace.rename handler (v2WorkspaceRename).
-        v2MainSync {
-            guard let ws = resolved.tabManager.tabs.first(where: { $0.id == resolved.workspaceId }) else {
-                return
-            }
-            var next = ws.metadata
-            for (k, maybeValue) in writes {
-                if let value = maybeValue {
-                    next[k] = value
-                } else {
-                    next.removeValue(forKey: k)
-                }
-            }
-            do {
-                try WorkspaceMetadataValidator.validateCapacity(after: next)
-            } catch let err as WorkspaceMetadataValidator.CapacityError {
-                capacityError = err
-                return
-            } catch {
-                // Unreachable: validateCapacity only throws CapacityError.
-                return
-            }
-            ws.metadata = next
-            resultMetadata = next
-        }
-
-        if let err = capacityError {
-            return .err(code: err.code, message: err.message, data: err.detail)
-        }
-
-        let windowId = v2ResolveWindowId(tabManager: resolved.tabManager)
-        return .ok([
-            "workspace_id": resolved.workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: resolved.workspaceId),
-            "window_id": v2OrNull(windowId?.uuidString),
-            "window_ref": v2Ref(kind: .window, uuid: windowId),
-            "metadata": resultMetadata
-        ])
-    }
-
-    private func v2WorkspaceGetMetadata(params: [String: Any]) -> V2CallResult {
-        let requestedKey = v2String(params, "key")
-        let requestedKeys = v2StringArray(params, "keys")
-
-        guard let resolved = v2ResolveWorkspaceForMetadata(params: params) else {
-            return .err(code: "not_found", message: "Workspace not found", data: nil)
-        }
-
-        var full: [String: String] = [:]
-        v2MainSync {
-            guard let ws = resolved.tabManager.tabs.first(where: { $0.id == resolved.workspaceId }) else {
-                return
-            }
-            full = ws.metadata
-        }
-
-        var metadataOut: [String: String] = full
-        if let filterKeys = requestedKeys {
-            metadataOut = [:]
-            for k in filterKeys {
-                if let v = full[k] { metadataOut[k] = v }
-            }
-        } else if let single = requestedKey {
-            metadataOut = [:]
-            if let v = full[single] { metadataOut[single] = v }
-        }
-
-        let windowId = v2ResolveWindowId(tabManager: resolved.tabManager)
-        var payload: [String: Any] = [
-            "workspace_id": resolved.workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: resolved.workspaceId),
-            "window_id": v2OrNull(windowId?.uuidString),
-            "window_ref": v2Ref(kind: .window, uuid: windowId),
-            "metadata": metadataOut
-        ]
-        if let single = requestedKey {
-            payload["key"] = single
-            payload["value"] = v2OrNull(metadataOut[single])
-        }
-        return .ok(payload)
-    }
-
-    private func v2WorkspaceClearMetadata(params: [String: Any]) -> V2CallResult {
-        let keys: [String]?
-        if params["keys"] == nil || params["keys"] is NSNull {
-            keys = nil
-        } else if let arr = v2StringArray(params, "keys") {
-            keys = arr
-        } else {
-            return .err(code: "invalid_keys_param", message: "keys must be an array of strings", data: nil)
-        }
-
-        // Validate key grammar off-main when a filter is provided; malformed
-        // keys could never have been written but reject them explicitly so
-        // callers see a consistent error shape.
-        if let keys {
-            for k in keys {
-                do {
-                    try WorkspaceMetadataValidator.validateKey(k)
-                } catch let err as WorkspaceMetadataValidator.ValidationError {
-                    return .err(code: err.code, message: err.message, data: err.detail)
-                } catch {
-                    return .err(code: "internal_error", message: "\(error)", data: nil)
-                }
-            }
-        }
-
-        guard let resolved = v2ResolveWorkspaceForMetadata(params: params) else {
-            return .err(code: "not_found", message: "Workspace not found", data: nil)
-        }
-
-        var resultMetadata: [String: String] = [:]
-        v2MainSync {
-            guard let ws = resolved.tabManager.tabs.first(where: { $0.id == resolved.workspaceId }) else {
-                return
-            }
-            if let keys {
-                var next = ws.metadata
-                for k in keys { next.removeValue(forKey: k) }
-                ws.metadata = next
-            } else {
-                ws.metadata = [:]
-            }
-            resultMetadata = ws.metadata
-        }
-
-        let windowId = v2ResolveWindowId(tabManager: resolved.tabManager)
-        return .ok([
-            "workspace_id": resolved.workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: resolved.workspaceId),
-            "window_id": v2OrNull(windowId?.uuidString),
-            "window_ref": v2Ref(kind: .window, uuid: windowId),
-            "metadata": resultMetadata
-        ])
-    }
-
-    private func v2WorkspaceAction(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let action = v2ActionKey(params) else {
-            return .err(code: "invalid_params", message: "Missing action", data: nil)
-        }
-
-        let supportedActions = [
-            "pin", "unpin", "rename", "clear_name",
-            "move_up", "move_down", "move_top",
-            "close_others", "close_above", "close_below",
-            "mark_read", "mark_unread"
-        ]
-
-        var result: V2CallResult = .err(code: "invalid_params", message: "Unknown workspace action", data: [
-            "action": action,
-            "supported_actions": supportedActions
-        ])
-
-        v2MainSync {
-            let requestedWorkspaceId = v2UUID(params, "workspace_id") ?? tabManager.selectedTabId
-            guard let workspaceId = requestedWorkspaceId,
-                  let workspace = tabManager.tabs.first(where: { $0.id == workspaceId }) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                return
-            }
-
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-
-            @MainActor
-            func closeWorkspaces(_ workspaces: [Workspace]) -> Int {
-                var closed = 0
-                for candidate in workspaces where candidate.id != workspace.id {
-                    let existedBefore = tabManager.tabs.contains(where: { $0.id == candidate.id })
-                    guard existedBefore else { continue }
-                    tabManager.closeWorkspace(candidate)
-                    if !tabManager.tabs.contains(where: { $0.id == candidate.id }) {
-                        closed += 1
-                    }
-                }
-                return closed
-            }
-
-            @MainActor
-            func finish(_ extras: [String: Any] = [:]) {
-                var payload: [String: Any] = [
-                    "action": action,
-                    "workspace_id": workspace.id.uuidString,
-                    "workspace_ref": v2Ref(kind: .workspace, uuid: workspace.id),
-                    "window_id": v2OrNull(windowId?.uuidString),
-                    "window_ref": v2Ref(kind: .window, uuid: windowId)
-                ]
-                for (key, value) in extras {
-                    payload[key] = value
-                }
-                result = .ok(payload)
-            }
-
-            switch action {
-            case "pin":
-                tabManager.setPinned(workspace, pinned: true)
-                finish(["pinned": true])
-
-            case "unpin":
-                tabManager.setPinned(workspace, pinned: false)
-                finish(["pinned": false])
-
-            case "rename":
-                guard let titleRaw = v2String(params, "title"),
-                      !titleRaw.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty else {
-                    result = .err(code: "invalid_params", message: "Missing or invalid title", data: nil)
-                    return
-                }
-                let title = titleRaw.trimmingCharacters(in: .whitespacesAndNewlines)
-                tabManager.setCustomTitle(tabId: workspace.id, title: title)
-                finish(["title": title])
-
-            case "clear_name":
-                tabManager.clearCustomTitle(tabId: workspace.id)
-                finish(["title": workspace.title])
-
-            case "move_up":
-                guard let currentIndex = tabManager.tabs.firstIndex(where: { $0.id == workspace.id }) else {
-                    result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                    return
-                }
-                _ = tabManager.reorderWorkspace(tabId: workspace.id, toIndex: max(currentIndex - 1, 0))
-                finish(["index": v2OrNull(tabManager.tabs.firstIndex(where: { $0.id == workspace.id }))])
-
-            case "move_down":
-                guard let currentIndex = tabManager.tabs.firstIndex(where: { $0.id == workspace.id }) else {
-                    result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                    return
-                }
-                _ = tabManager.reorderWorkspace(tabId: workspace.id, toIndex: min(currentIndex + 1, tabManager.tabs.count - 1))
-                finish(["index": v2OrNull(tabManager.tabs.firstIndex(where: { $0.id == workspace.id }))])
-
-            case "move_top":
-                tabManager.moveTabToTop(workspace.id)
-                finish(["index": v2OrNull(tabManager.tabs.firstIndex(where: { $0.id == workspace.id }))])
-
-            case "close_others":
-                let candidates = tabManager.tabs.filter { $0.id != workspace.id && !$0.isPinned }
-                let closed = closeWorkspaces(candidates)
-                finish(["closed": closed])
-
-            case "close_above":
-                guard let index = tabManager.tabs.firstIndex(where: { $0.id == workspace.id }) else {
-                    result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                    return
-                }
-                let candidates = Array(tabManager.tabs.prefix(index)).filter { !$0.isPinned }
-                let closed = closeWorkspaces(candidates)
-                finish(["closed": closed])
-
-            case "close_below":
-                guard let index = tabManager.tabs.firstIndex(where: { $0.id == workspace.id }) else {
-                    result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                    return
-                }
-                let candidates: [Workspace]
-                if index + 1 < tabManager.tabs.count {
-                    candidates = Array(tabManager.tabs.suffix(from: index + 1)).filter { !$0.isPinned }
-                } else {
-                    candidates = []
-                }
-                let closed = closeWorkspaces(candidates)
-                finish(["closed": closed])
-
-            case "mark_read":
-                AppDelegate.shared?.notificationStore?.markRead(forTabId: workspace.id)
-                finish()
-
-            case "mark_unread":
-                AppDelegate.shared?.notificationStore?.markUnread(forTabId: workspace.id)
-                finish()
-
-            default:
-                result = .err(code: "invalid_params", message: "Unknown workspace action", data: [
-                    "action": action,
-                    "supported_actions": supportedActions
-                ])
-            }
-        }
-
-        return result
-    }
 
 
     // MARK: - V2 Surface Methods
@@ -7579,254 +6158,7 @@ class TerminalController {
 
     // MARK: - V2 Pane Methods
 
-    private func v2PaneList(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
 
-        var payload: [String: Any]?
-        v2MainSync {
-            guard let ws = v2ResolveWorkspace(params: params, tabManager: tabManager) else { return }
-
-            let focusedPaneId = ws.bonsplitController.focusedPaneId
-            let panes: [[String: Any]] = ws.bonsplitController.allPaneIds.enumerated().map { index, paneId in
-                let tabs = ws.bonsplitController.tabs(inPane: paneId)
-                let surfaceUUIDs: [UUID] = tabs.compactMap { ws.panelIdFromSurfaceId($0.id) }
-                let selectedTab = ws.bonsplitController.selectedTab(inPane: paneId)
-                let selectedSurfaceUUID = selectedTab.flatMap { ws.panelIdFromSurfaceId($0.id) }
-                return [
-                    "id": paneId.id.uuidString,
-                    "ref": v2Ref(kind: .pane, uuid: paneId.id),
-                    "index": index,
-                    "focused": paneId == focusedPaneId,
-                    "surface_ids": surfaceUUIDs.map { $0.uuidString },
-                    "surface_refs": surfaceUUIDs.map { v2Ref(kind: .surface, uuid: $0) },
-                    "selected_surface_id": v2OrNull(selectedSurfaceUUID?.uuidString),
-                    "selected_surface_ref": v2Ref(kind: .surface, uuid: selectedSurfaceUUID),
-                    "surface_count": surfaceUUIDs.count
-                ]
-            }
-
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            payload = [
-                "workspace_id": ws.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id),
-                "panes": panes,
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId)
-            ]
-        }
-
-        guard let payload else {
-            return .err(code: "not_found", message: "Workspace not found", data: nil)
-        }
-        return .ok(payload)
-    }
-    private func v2PaneFocus(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let paneUUID = v2UUID(params, "pane_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid pane_id", data: nil)
-        }
-
-        var result: V2CallResult = .err(code: "not_found", message: "Pane not found", data: ["pane_id": paneUUID.uuidString])
-        v2MainSync {
-            guard let ws = v2ResolveWorkspace(params: params, tabManager: tabManager) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                return
-            }
-            guard let paneId = ws.bonsplitController.allPaneIds.first(where: { $0.id == paneUUID }) else {
-                result = .err(code: "not_found", message: "Pane not found", data: ["pane_id": paneUUID.uuidString])
-                return
-            }
-            if let windowId = v2ResolveWindowId(tabManager: tabManager) {
-                _ = AppDelegate.shared?.focusMainWindow(windowId: windowId)
-                setActiveTabManager(tabManager)
-            }
-            if tabManager.selectedTabId != ws.id {
-                tabManager.selectWorkspace(ws)
-            }
-            ws.bonsplitController.focusPane(paneId)
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            result = .ok(["window_id": v2OrNull(windowId?.uuidString), "window_ref": v2Ref(kind: .window, uuid: windowId), "workspace_id": ws.id.uuidString, "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id), "pane_id": paneId.id.uuidString, "pane_ref": v2Ref(kind: .pane, uuid: paneId.id)])
-        }
-        return result
-    }
-
-    private func v2PaneSurfaces(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-
-        var payload: [String: Any]?
-        v2MainSync {
-            guard let ws = v2ResolveWorkspace(params: params, tabManager: tabManager) else { return }
-
-            let paneUUID = v2UUID(params, "pane_id")
-            let paneId: PaneID? = {
-                if let paneUUID {
-                    return ws.bonsplitController.allPaneIds.first(where: { $0.id == paneUUID })
-                }
-                return ws.bonsplitController.focusedPaneId
-            }()
-            guard let paneId else { return }
-
-            let selectedTab = ws.bonsplitController.selectedTab(inPane: paneId)
-            let tabs = ws.bonsplitController.tabs(inPane: paneId)
-
-            let surfaces: [[String: Any]] = tabs.enumerated().map { index, tab in
-                let panelId = ws.panelIdFromSurfaceId(tab.id)
-                let panel = panelId.flatMap { ws.panels[$0] }
-                return [
-                    "id": v2OrNull(panelId?.uuidString),
-                    "ref": v2Ref(kind: .surface, uuid: panelId),
-                    "index": index,
-                    "title": tab.title,
-                    "type": v2OrNull(panel?.panelType.rawValue),
-                    "selected": tab.id == selectedTab?.id
-                ]
-            }
-
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            payload = [
-                "workspace_id": ws.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id),
-                "pane_id": paneId.id.uuidString,
-                "pane_ref": v2Ref(kind: .pane, uuid: paneId.id),
-                "surfaces": surfaces,
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId)
-            ]
-        }
-
-        guard let payload else {
-            return .err(code: "not_found", message: "Pane or workspace not found", data: nil)
-        }
-        return .ok(payload)
-    }
-    private func v2PaneCreate(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let directionStr = v2String(params, "direction"),
-              let direction = parseSplitDirection(directionStr) else {
-            return .err(code: "invalid_params", message: "Missing or invalid direction (left|right|up|down)", data: nil)
-        }
-
-        let panelType = v2PanelType(params, "type") ?? .terminal
-        if let denial = v2SurfaceTypeDenial(panelType) { return denial }
-        let urlStr = v2String(params, "url")
-        let url = urlStr.flatMap { URL(string: $0) }
-        let filePath = v2String(params, "file")
-        let titleSeed = v2String(params, "title")
-
-        // Validate and resolve markdown file path
-        var resolvedMarkdownPath: String?
-        if panelType == .markdown {
-            if let err = v2ValidateMarkdownPath(filePath, context: "pane", resolved: &resolvedMarkdownPath) {
-                return err
-            }
-        }
-
-        // Validate the optional --cwd override server-side. Only meaningful for
-        // terminal panes (browser/markdown have no shell), but validate
-        // regardless so a bad path is rejected rather than silently ignored.
-        var cwdOverride: String?
-        if let err = v2ResolveCwdParam(params, resolved: &cwdOverride) {
-            return err
-        }
-
-        let force = splitForceFlag(params)
-
-        var result: V2CallResult = .err(code: "internal_error", message: "Failed to create pane", data: nil)
-        guard v2MainSyncWithDeadline({
-            guard let ws = self.v2ResolveWorkspace(params: params, tabManager: tabManager) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                return
-            }
-            self.v2MaybeFocusWindow(for: tabManager)
-            self.v2MaybeSelectWorkspace(tabManager, workspace: ws)
-            guard let focusedPanelId = ws.focusedPanelId else {
-                result = .err(code: "not_found", message: "No focused surface to split", data: nil)
-                return
-            }
-
-            let plan = self.planSizeAwareSplit(
-                ws: ws,
-                sourcePanelId: focusedPanelId,
-                requested: direction,
-                newIsTerminal: panelType == .terminal,
-                force: force
-            )
-
-            var becameTab = false
-            var appliedDirection = direction
-            var warningText: String?
-            var targetPaneForTab: PaneID?
-            var newPanelId: UUID?
-
-            switch plan {
-            case .refuse(let message, let data):
-                result = .err(code: "pane_too_small", message: message, data: data)
-                return
-
-            case .tab(let paneId, let warning):
-                becameTab = true
-                warningText = warning
-                targetPaneForTab = paneId
-                switch panelType {
-                case .browser:
-                    newPanelId = ws.newBrowserSurface(inPane: paneId, url: url, focus: self.v2FocusAllowed())?.id
-                case .markdown:
-                    newPanelId = ws.newMarkdownSurface(inPane: paneId, filePath: resolvedMarkdownPath!, focus: self.v2FocusAllowed())?.id
-                case .terminal:
-                    newPanelId = ws.newTerminalSurface(inPane: paneId, focus: self.v2FocusAllowed(), workingDirectory: cwdOverride)?.id
-                }
-
-            case .split(let actualDirection, _, let warning):
-                appliedDirection = actualDirection
-                warningText = warning
-                let orientation = actualDirection.orientation
-                let insertFirst = actualDirection.insertFirst
-                switch panelType {
-                case .browser:
-                    newPanelId = ws.newBrowserSplit(from: focusedPanelId, orientation: orientation, insertFirst: insertFirst, url: url, focus: self.v2FocusAllowed())?.id
-                case .markdown:
-                    newPanelId = ws.newMarkdownSplit(from: focusedPanelId, orientation: orientation, insertFirst: insertFirst, filePath: resolvedMarkdownPath!, focus: self.v2FocusAllowed())?.id
-                case .terminal:
-                    newPanelId = ws.newTerminalSplit(from: focusedPanelId, orientation: orientation, insertFirst: insertFirst, focus: self.v2FocusAllowed(), workingDirectory: cwdOverride)?.id
-                }
-            }
-
-            guard let createdPanelId = newPanelId else {
-                result = .err(code: "internal_error", message: "Failed to create pane", data: nil)
-                return
-            }
-            let paneUUID = becameTab ? targetPaneForTab?.id : ws.paneId(forPanelId: createdPanelId)?.id
-            // Seed pane title atomic with the pane id becoming valid: the
-            // caller observes the pane (via the response) only after the seed
-            // is in the store.
-            self.v2SeedPaneTitle(workspaceId: ws.id, paneUUID: paneUUID, title: titleSeed)
-            let windowId = self.v2ResolveWindowId(tabManager: tabManager)
-            var ok: [String: Any] = [
-                "window_id": self.v2OrNull(windowId?.uuidString),
-                "window_ref": self.v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": ws.id.uuidString,
-                "workspace_ref": self.v2Ref(kind: .workspace, uuid: ws.id),
-                "pane_id": self.v2OrNull(paneUUID?.uuidString),
-                "pane_ref": self.v2Ref(kind: .pane, uuid: paneUUID),
-                "surface_id": createdPanelId.uuidString,
-                "surface_ref": self.v2Ref(kind: .surface, uuid: createdPanelId),
-                "type": panelType.rawValue
-            ]
-            self.annotateSizeOutcome(&ok, requested: direction, applied: appliedDirection, becameTab: becameTab, warning: warningText)
-            result = .ok(ok)
-        }) != nil else {
-            return .err(code: "main_thread_timeout", message: "main thread did not respond within deadline", data: nil)
-        }
-        return result
-    }
 
     private enum V2PaneResizeDirection: String {
         case left
@@ -7873,386 +6205,11 @@ class TerminalController {
         let bounds: CGRect
     }
 
-    private func v2PaneResizeCollectCandidates(
-        node: ExternalTreeNode,
-        targetPaneId: String,
-        candidates: inout [V2PaneResizeCandidate]
-    ) -> V2PaneResizeTrace {
-        switch node {
-        case .pane(let pane):
-            let bounds = CGRect(
-                x: pane.frame.x,
-                y: pane.frame.y,
-                width: pane.frame.width,
-                height: pane.frame.height
-            )
-            return V2PaneResizeTrace(containsTarget: pane.id == targetPaneId, bounds: bounds)
 
-        case .split(let split):
-            let first = v2PaneResizeCollectCandidates(
-                node: split.first,
-                targetPaneId: targetPaneId,
-                candidates: &candidates
-            )
-            let second = v2PaneResizeCollectCandidates(
-                node: split.second,
-                targetPaneId: targetPaneId,
-                candidates: &candidates
-            )
 
-            let combinedBounds = first.bounds.union(second.bounds)
-            let containsTarget = first.containsTarget || second.containsTarget
 
-            if containsTarget,
-               let splitUUID = UUID(uuidString: split.id) {
-                let orientation = split.orientation.lowercased()
-                let axisPixels: CGFloat = orientation == "horizontal"
-                    ? combinedBounds.width
-                    : combinedBounds.height
-                candidates.append(V2PaneResizeCandidate(
-                    splitId: splitUUID,
-                    orientation: orientation,
-                    paneInFirstChild: first.containsTarget,
-                    dividerPosition: CGFloat(split.dividerPosition),
-                    axisPixels: max(axisPixels, 1)
-                ))
-            }
 
-            return V2PaneResizeTrace(containsTarget: containsTarget, bounds: combinedBounds)
-        }
-    }
 
-    private func v2PaneResize(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-
-        let directionRaw = (v2String(params, "direction") ?? "").lowercased()
-        let amount = v2Int(params, "amount") ?? 1
-        guard let direction = V2PaneResizeDirection(rawValue: directionRaw), amount > 0 else {
-            return .err(code: "invalid_params", message: "direction must be one of left|right|up|down and amount must be > 0", data: nil)
-        }
-
-        var result: V2CallResult = .err(code: "internal_error", message: "Failed to resize pane", data: nil)
-        v2MainSync {
-            guard let ws = v2ResolveWorkspace(params: params, tabManager: tabManager) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                return
-            }
-
-            let paneUUID = v2UUID(params, "pane_id") ?? ws.bonsplitController.focusedPaneId?.id
-            guard let paneUUID else {
-                result = .err(code: "not_found", message: "No focused pane", data: nil)
-                return
-            }
-            guard ws.bonsplitController.allPaneIds.contains(where: { $0.id == paneUUID }) else {
-                result = .err(code: "not_found", message: "Pane not found", data: ["pane_id": paneUUID.uuidString])
-                return
-            }
-
-            let tree = ws.bonsplitController.treeSnapshot()
-            var candidates: [V2PaneResizeCandidate] = []
-            let trace = v2PaneResizeCollectCandidates(
-                node: tree,
-                targetPaneId: paneUUID.uuidString,
-                candidates: &candidates
-            )
-            guard trace.containsTarget else {
-                result = .err(code: "not_found", message: "Pane not found in split tree", data: ["pane_id": paneUUID.uuidString])
-                return
-            }
-
-            let orientationMatches = candidates.filter { $0.orientation == direction.splitOrientation }
-            guard !orientationMatches.isEmpty else {
-                result = .err(
-                    code: "invalid_state",
-                    message: "No \(direction.splitOrientation) split ancestor for pane",
-                    data: ["pane_id": paneUUID.uuidString, "direction": direction.rawValue]
-                )
-                return
-            }
-
-            guard let candidate = orientationMatches.first(where: { $0.paneInFirstChild == direction.requiresPaneInFirstChild }) else {
-                result = .err(
-                    code: "invalid_state",
-                    message: "Pane has no adjacent border in direction \(direction.rawValue)",
-                    data: ["pane_id": paneUUID.uuidString, "direction": direction.rawValue]
-                )
-                return
-            }
-
-            let delta = CGFloat(amount) / candidate.axisPixels
-            let requested = candidate.dividerPosition + (direction.dividerDeltaSign * delta)
-            let clamped = min(max(requested, 0.1), 0.9)
-            guard ws.bonsplitController.setDividerPosition(clamped, forSplit: candidate.splitId, fromExternal: true) else {
-                result = .err(
-                    code: "internal_error",
-                    message: "Failed to set split divider position",
-                    data: ["split_id": candidate.splitId.uuidString]
-                )
-                return
-            }
-
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            result = .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": ws.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id),
-                "pane_id": paneUUID.uuidString,
-                "pane_ref": v2Ref(kind: .pane, uuid: paneUUID),
-                "split_id": candidate.splitId.uuidString,
-                "direction": direction.rawValue,
-                "amount": amount,
-                "old_divider_position": candidate.dividerPosition,
-                "new_divider_position": clamped
-            ])
-        }
-        return result
-    }
-
-    private func v2PaneSwap(params: [String: Any]) -> V2CallResult {
-        guard let sourcePaneUUID = v2UUID(params, "pane_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid pane_id", data: nil)
-        }
-        guard let targetPaneUUID = v2UUID(params, "target_pane_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid target_pane_id", data: nil)
-        }
-        if sourcePaneUUID == targetPaneUUID {
-            return .err(code: "invalid_params", message: "pane_id and target_pane_id must be different", data: nil)
-        }
-        let focus = v2FocusAllowed(requested: v2Bool(params, "focus") ?? true)
-
-        var result: V2CallResult = .err(code: "internal_error", message: "Failed to swap panes", data: nil)
-        v2MainSync {
-            guard let located = v2LocatePane(sourcePaneUUID) else {
-                result = .err(code: "not_found", message: "Source pane not found", data: ["pane_id": sourcePaneUUID.uuidString])
-                return
-            }
-            guard let targetPane = located.workspace.bonsplitController.allPaneIds.first(where: { $0.id == targetPaneUUID }) else {
-                result = .err(code: "not_found", message: "Target pane not found in source workspace", data: ["target_pane_id": targetPaneUUID.uuidString])
-                return
-            }
-            let workspace = located.workspace
-            let sourcePane = located.paneId
-
-            guard let selectedSourceTab = workspace.bonsplitController.selectedTab(inPane: sourcePane),
-                  let selectedTargetTab = workspace.bonsplitController.selectedTab(inPane: targetPane),
-                  let sourceSurfaceId = workspace.panelIdFromSurfaceId(selectedSourceTab.id),
-                  let targetSurfaceId = workspace.panelIdFromSurfaceId(selectedTargetTab.id) else {
-                result = .err(code: "invalid_state", message: "Both panes must have a selected surface", data: nil)
-                return
-            }
-
-            // Keep pane identities stable during swap when one side has a single surface.
-            var sourcePlaceholder: UUID?
-            var targetPlaceholder: UUID?
-            if workspace.bonsplitController.tabs(inPane: sourcePane).count <= 1 {
-                sourcePlaceholder = workspace.newTerminalSurface(inPane: sourcePane, focus: false)?.id
-                if sourcePlaceholder == nil {
-                    result = .err(code: "internal_error", message: "Failed to create source placeholder surface", data: nil)
-                    return
-                }
-            }
-            if workspace.bonsplitController.tabs(inPane: targetPane).count <= 1 {
-                targetPlaceholder = workspace.newTerminalSurface(inPane: targetPane, focus: false)?.id
-                if targetPlaceholder == nil {
-                    result = .err(code: "internal_error", message: "Failed to create target placeholder surface", data: nil)
-                    return
-                }
-            }
-
-            guard workspace.moveSurface(panelId: sourceSurfaceId, toPane: targetPane, focus: false) else {
-                result = .err(code: "internal_error", message: "Failed moving source surface into target pane", data: nil)
-                return
-            }
-            guard workspace.moveSurface(panelId: targetSurfaceId, toPane: sourcePane, focus: false) else {
-                result = .err(code: "internal_error", message: "Failed moving target surface into source pane", data: nil)
-                return
-            }
-
-            if let sourcePlaceholder {
-                _ = workspace.closePanel(sourcePlaceholder, force: true)
-            }
-            if let targetPlaceholder {
-                _ = workspace.closePanel(targetPlaceholder, force: true)
-            }
-
-            if focus {
-                workspace.bonsplitController.focusPane(targetPane)
-            }
-            let windowId = located.windowId
-            result = .ok([
-                "window_id": windowId.uuidString,
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": workspace.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: workspace.id),
-                "pane_id": sourcePane.id.uuidString,
-                "pane_ref": v2Ref(kind: .pane, uuid: sourcePane.id),
-                "target_pane_id": targetPane.id.uuidString,
-                "target_pane_ref": v2Ref(kind: .pane, uuid: targetPane.id),
-                "source_surface_id": sourceSurfaceId.uuidString,
-                "source_surface_ref": v2Ref(kind: .surface, uuid: sourceSurfaceId),
-                "target_surface_id": targetSurfaceId.uuidString,
-                "target_surface_ref": v2Ref(kind: .surface, uuid: targetSurfaceId)
-            ])
-        }
-        return result
-    }
-
-    private func v2PaneBreak(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        let focus = v2FocusAllowed(requested: v2Bool(params, "focus") ?? true)
-
-        var result: V2CallResult = .err(code: "internal_error", message: "Failed to break pane", data: nil)
-        v2MainSync {
-            guard let sourceWorkspace = v2ResolveWorkspace(params: params, tabManager: tabManager) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                return
-            }
-
-            let sourcePaneUUID = v2UUID(params, "pane_id")
-            let sourcePane: PaneID? = {
-                if let sourcePaneUUID {
-                    return sourceWorkspace.bonsplitController.allPaneIds.first(where: { $0.id == sourcePaneUUID })
-                }
-                return sourceWorkspace.bonsplitController.focusedPaneId
-            }()
-
-            let surfaceId: UUID? = {
-                if let explicitSurface = v2UUID(params, "surface_id") { return explicitSurface }
-                if let sourcePane,
-                   let selected = sourceWorkspace.bonsplitController.selectedTab(inPane: sourcePane) {
-                    return sourceWorkspace.panelIdFromSurfaceId(selected.id)
-                }
-                return sourceWorkspace.focusedPanelId
-            }()
-            guard let surfaceId else {
-                result = .err(code: "not_found", message: "No source surface to break", data: nil)
-                return
-            }
-            guard sourceWorkspace.panels[surfaceId] != nil else {
-                result = .err(code: "not_found", message: "Surface not found", data: ["surface_id": surfaceId.uuidString])
-                return
-            }
-            let sourceIndex = sourceWorkspace.indexInPane(forPanelId: surfaceId)
-            let sourcePaneForRollback = sourceWorkspace.paneId(forPanelId: surfaceId)
-
-            guard let detached = sourceWorkspace.detachSurface(panelId: surfaceId) else {
-                result = .err(code: "internal_error", message: "Failed to detach source surface", data: nil)
-                return
-            }
-
-            let destinationWorkspace = tabManager.addWorkspace(select: focus)
-            guard let destinationPane = destinationWorkspace.bonsplitController.focusedPaneId
-                ?? destinationWorkspace.bonsplitController.allPaneIds.first else {
-                if let sourcePaneForRollback {
-                    _ = sourceWorkspace.attachDetachedSurface(
-                        detached,
-                        inPane: sourcePaneForRollback,
-                        atIndex: sourceIndex,
-                        focus: true
-                    )
-                }
-                result = .err(code: "internal_error", message: "Destination workspace has no pane", data: nil)
-                return
-            }
-
-            guard destinationWorkspace.attachDetachedSurface(detached, inPane: destinationPane, focus: focus) != nil else {
-                if let sourcePaneForRollback {
-                    _ = sourceWorkspace.attachDetachedSurface(
-                        detached,
-                        inPane: sourcePaneForRollback,
-                        atIndex: sourceIndex,
-                        focus: true
-                    )
-                }
-                result = .err(code: "internal_error", message: "Failed to attach surface to new workspace", data: nil)
-                return
-            }
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            result = .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": destinationWorkspace.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: destinationWorkspace.id),
-                "pane_id": destinationPane.id.uuidString,
-                "pane_ref": v2Ref(kind: .pane, uuid: destinationPane.id),
-                "surface_id": surfaceId.uuidString,
-                "surface_ref": v2Ref(kind: .surface, uuid: surfaceId)
-            ])
-        }
-        return result
-    }
-
-    private func v2PaneJoin(params: [String: Any]) -> V2CallResult {
-        guard let targetPaneUUID = v2UUID(params, "target_pane_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid target_pane_id", data: nil)
-        }
-
-        var surfaceId = v2UUID(params, "surface_id")
-        if surfaceId == nil, let sourcePaneUUID = v2UUID(params, "pane_id") {
-            guard let sourceLocated = v2LocatePane(sourcePaneUUID),
-                  let selected = sourceLocated.workspace.bonsplitController.selectedTab(inPane: sourceLocated.paneId),
-                  let selectedSurface = sourceLocated.workspace.panelIdFromSurfaceId(selected.id) else {
-                return .err(code: "not_found", message: "Unable to resolve selected surface in source pane", data: [
-                    "pane_id": sourcePaneUUID.uuidString
-                ])
-            }
-            surfaceId = selectedSurface
-        }
-        guard let surfaceId else {
-            return .err(code: "invalid_params", message: "Missing surface_id (or pane_id with selected surface)", data: nil)
-        }
-
-        var moveParams: [String: Any] = [
-            "surface_id": surfaceId.uuidString,
-            "pane_id": targetPaneUUID.uuidString
-        ]
-        if let focus = v2Bool(params, "focus") {
-            moveParams["focus"] = focus
-        }
-        return v2SurfaceMove(params: moveParams)
-    }
-
-    private func v2PaneLast(params: [String: Any]) -> V2CallResult {
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-
-        var result: V2CallResult = .err(code: "not_found", message: "No alternate pane available", data: nil)
-        v2MainSync {
-            guard let ws = v2ResolveWorkspace(params: params, tabManager: tabManager) else {
-                result = .err(code: "not_found", message: "Workspace not found", data: nil)
-                return
-            }
-            guard let focused = ws.bonsplitController.focusedPaneId else {
-                result = .err(code: "not_found", message: "No focused pane", data: nil)
-                return
-            }
-            guard let target = ws.bonsplitController.allPaneIds.first(where: { $0.id != focused.id }) else {
-                result = .err(code: "not_found", message: "No alternate pane available", data: nil)
-                return
-            }
-
-            ws.bonsplitController.focusPane(target)
-            let selectedSurfaceId = ws.bonsplitController.selectedTab(inPane: target).flatMap { ws.panelIdFromSurfaceId($0.id) }
-            let windowId = v2ResolveWindowId(tabManager: tabManager)
-            result = .ok([
-                "window_id": v2OrNull(windowId?.uuidString),
-                "window_ref": v2Ref(kind: .window, uuid: windowId),
-                "workspace_id": ws.id.uuidString,
-                "workspace_ref": v2Ref(kind: .workspace, uuid: ws.id),
-                "pane_id": target.id.uuidString,
-                "pane_ref": v2Ref(kind: .pane, uuid: target.id),
-                "surface_id": v2OrNull(selectedSurfaceId?.uuidString),
-                "surface_ref": v2Ref(kind: .surface, uuid: selectedSurfaceId)
-            ])
-        }
-        return result
-    }
 
     // MARK: - V2 Pane Metadata (CMUX-11 Phase 2)
 
@@ -8284,160 +6241,8 @@ class TerminalController {
         }
     }
 
-    private func v2PaneSetMetadata(params: [String: Any]) -> V2CallResult {
-        guard let metadataObj = params["metadata"] as? [String: Any] else {
-            return .err(code: "invalid_json", message: "metadata must be a JSON object", data: nil)
-        }
 
-        let modeStr = (v2String(params, "mode") ?? "merge").lowercased()
-        guard let mode = SurfaceMetadataStore.WriteMode(rawValue: modeStr) else {
-            return .err(code: "invalid_mode", message: "mode must be 'merge' or 'replace'", data: nil)
-        }
 
-        let sourceStr = (v2String(params, "source") ?? "explicit").lowercased()
-        guard let source = MetadataSource(rawValue: sourceStr) else {
-            return .err(code: "invalid_source", message: "source must be one of: explicit, declare, osc, heuristic", data: nil)
-        }
-        // C11-104 v2 (B5a) — `derived` is reserved for c11-internal
-        // writers. External socket/CLI clients are rejected. Without
-        // this gate, an agent could write `source=derived` and claim
-        // their values are system-computed, nullifying the meaning
-        // of the precedence tier.
-        //
-        // (C11-106 AC16) Logic moved to SocketMetadataSourceValidator
-        // so the rejection contract is exercised in
-        // c11Tests/SocketDerivedSourceRejectionTests.swift without
-        // standing up a full socket frame loop.
-        if let rejection = SocketMetadataSourceValidator.externalRejectionMessage(for: source) {
-            return .err(code: rejection.code, message: rejection.message, data: nil)
-        }
-
-        guard let resolved = v2ResolvePaneForMetadata(params: params) else {
-            return .err(code: "pane_not_found", message: "Pane not found", data: nil)
-        }
-
-        do {
-            let result = try PaneMetadataStore.shared.setMetadata(
-                workspaceId: resolved.workspaceId,
-                paneId: resolved.paneId,
-                partial: metadataObj,
-                mode: mode,
-                source: source
-            )
-            return .ok(buildPaneMetadataOkPayload(
-                workspaceId: resolved.workspaceId,
-                paneId: resolved.paneId,
-                tabManager: resolved.tabManager,
-                result: result,
-                includePriorValues: true
-            ))
-        } catch let err as SurfaceMetadataStore.WriteError {
-            return .err(code: err.code, message: err.message, data: err.detailData)
-        } catch {
-            return .err(code: "internal_error", message: "\(error)", data: nil)
-        }
-    }
-
-    private func v2PaneGetMetadata(params: [String: Any]) -> V2CallResult {
-        let keys: [String]?
-        if params["keys"] is NSNull || params["keys"] == nil {
-            keys = nil
-        } else if let arr = v2StringArray(params, "keys") {
-            keys = arr
-        } else {
-            return .err(code: "invalid_keys_param", message: "keys must be an array of strings", data: nil)
-        }
-
-        let includeSources = v2Bool(params, "include_sources") ?? false
-
-        guard let resolved = v2ResolvePaneForMetadata(params: params) else {
-            return .err(code: "pane_not_found", message: "Pane not found", data: nil)
-        }
-
-        let (fullMetadata, fullSources) = PaneMetadataStore.shared.getMetadata(
-            workspaceId: resolved.workspaceId,
-            paneId: resolved.paneId
-        )
-
-        var metadataOut: [String: Any] = fullMetadata
-        var sourcesOut: [String: [String: Any]] = fullSources
-        if let filterKeys = keys {
-            metadataOut = [:]
-            sourcesOut = [:]
-            for k in filterKeys {
-                if let v = fullMetadata[k] { metadataOut[k] = v }
-                if let s = fullSources[k] { sourcesOut[k] = s }
-            }
-        }
-
-        let windowId = v2ResolveWindowId(tabManager: resolved.tabManager)
-        var payload: [String: Any] = [
-            "workspace_id": resolved.workspaceId.uuidString,
-            "workspace_ref": v2Ref(kind: .workspace, uuid: resolved.workspaceId),
-            "pane_id": resolved.paneId.uuidString,
-            "pane_ref": v2Ref(kind: .pane, uuid: resolved.paneId),
-            "window_id": v2OrNull(windowId?.uuidString),
-            "window_ref": v2Ref(kind: .window, uuid: windowId),
-            "metadata": metadataOut
-        ]
-        if includeSources {
-            payload["metadata_sources"] = sourcesOut
-        }
-        return .ok(payload)
-    }
-
-    private func v2PaneClearMetadata(params: [String: Any]) -> V2CallResult {
-        let keys: [String]?
-        if params["keys"] == nil || params["keys"] is NSNull {
-            keys = nil
-        } else if let arr = v2StringArray(params, "keys") {
-            keys = arr
-        } else {
-            return .err(code: "invalid_keys_param", message: "keys must be an array of strings", data: nil)
-        }
-
-        let sourceStr = (v2String(params, "source") ?? "explicit").lowercased()
-        guard let source = MetadataSource(rawValue: sourceStr) else {
-            return .err(code: "invalid_source", message: "source must be one of: explicit, declare, osc, heuristic", data: nil)
-        }
-        // C11-104 v2 (B5a) — `derived` is reserved for c11-internal
-        // writers. External socket/CLI clients are rejected. Without
-        // this gate, an agent could write `source=derived` and claim
-        // their values are system-computed, nullifying the meaning
-        // of the precedence tier.
-        //
-        // (C11-106 AC16) Logic moved to SocketMetadataSourceValidator
-        // so the rejection contract is exercised in
-        // c11Tests/SocketDerivedSourceRejectionTests.swift without
-        // standing up a full socket frame loop.
-        if let rejection = SocketMetadataSourceValidator.externalRejectionMessage(for: source) {
-            return .err(code: rejection.code, message: rejection.message, data: nil)
-        }
-
-        guard let resolved = v2ResolvePaneForMetadata(params: params) else {
-            return .err(code: "pane_not_found", message: "Pane not found", data: nil)
-        }
-
-        do {
-            let result = try PaneMetadataStore.shared.clearMetadata(
-                workspaceId: resolved.workspaceId,
-                paneId: resolved.paneId,
-                keys: keys,
-                source: source
-            )
-            return .ok(buildPaneMetadataOkPayload(
-                workspaceId: resolved.workspaceId,
-                paneId: resolved.paneId,
-                tabManager: resolved.tabManager,
-                result: result,
-                includePriorValues: false
-            ))
-        } catch let err as SurfaceMetadataStore.WriteError {
-            return .err(code: err.code, message: err.message, data: err.detailData)
-        } catch {
-            return .err(code: "internal_error", message: "\(error)", data: nil)
-        }
-    }
 
     private func buildPaneMetadataOkPayload(
         workspaceId: UUID,
@@ -8469,181 +6274,8 @@ class TerminalController {
         return payload
     }
 
-    /// Seed `PaneMetadataStore` with `{title: <value>}` at pane creation time.
-    /// Called from `v2PaneCreate` / `v2SurfaceSplit` when the caller passes a
-    /// `title` parameter. Atomic with the pane id becoming valid — agents can
-    /// assume the seed is present before the RPC response returns.
-    private func v2SeedPaneTitle(
-        workspaceId: UUID,
-        paneUUID: UUID?,
-        title: String?
-    ) {
-        guard let paneUUID, let title, !title.isEmpty else { return }
-        do {
-            _ = try PaneMetadataStore.shared.setMetadata(
-                workspaceId: workspaceId,
-                paneId: paneUUID,
-                partial: [MetadataKey.title: title],
-                mode: .merge,
-                source: .explicit
-            )
-        } catch {
-            // Seeding is best-effort: the pane is already created and the
-            // caller can retry via pane.set_metadata. Log and continue.
-            #if DEBUG
-            dlog("pane.title_seed.failed pane=\(paneUUID.uuidString) err=\(error)")
-            #endif
-        }
-    }
 
-    /// Socket-triggered pane confirmation. Presents a .confirm interaction on the
-    /// panel identified by panel_id (surface UUID) and blocks the socket call
-    /// until the user accepts, cancels, or the dialog is dismissed (panel torn
-    /// down mid-prompt). Per CLAUDE.md's socket threading policy, pane.confirm is
-    /// an explicit focus-intent command — main-thread UI mutation is allowed.
-    ///
-    /// Params:
-    ///   - panel_id (string, required): surface UUID of the target panel
-    ///   - title (string, required): card title
-    ///   - message (string, optional): card informative text
-    ///   - role (string, optional): "destructive" | "standard" (default: standard)
-    ///   - timeout (number, optional): max seconds to wait; clamped to [0, 300].
-    ///     Omitting defaults to 300 seconds — the server-side ceiling prevents a
-    ///     malformed or malicious caller from blocking a socket worker with an
-    ///     indefinite wait. Caller-supplied values > 300s are silently clamped.
-    ///   - confirm_label (string, optional): override for the confirm button label
-    ///   - cancel_label (string, optional): override for the cancel button label
-    ///
-    /// Result: { "result": "ok" | "cancel" | "dismissed" }
-    ///
-    /// Errors:
-    ///   - "unavailable" — no TabManager
-    ///   - "invalid_params" — missing panel_id or title
-    ///   - "unknown_panel" — panel_id doesn't resolve to a panel in any workspace
-    private func v2PaneConfirm(params: [String: Any]) -> V2CallResult {
-        // Document and enforce the off-main contract: v2PaneConfirm blocks the
-        // calling thread on a DispatchSemaphore, so running this on main would
-        // hang the UI (v2MainSync falls through to direct execution when
-        // already on main, which would then dead-wait). Socket dispatch runs
-        // off-main via Thread.detachNewThread; this guard catches regressions.
-        assert(!Thread.isMainThread, "v2PaneConfirm must not be called on the main thread")
 
-        guard let tabManager = v2ResolveTabManager(params: params) else {
-            return .err(code: "unavailable", message: "TabManager not available", data: nil)
-        }
-        guard let panelId = v2UUID(params, "panel_id") else {
-            return .err(code: "invalid_params", message: "Missing or invalid panel_id", data: nil)
-        }
-        guard let title = v2String(params, "title"), !title.isEmpty else {
-            return .err(code: "invalid_params", message: "Missing or empty title", data: nil)
-        }
-        let message = (v2String(params, "message") ?? "")
-        let roleStr = v2String(params, "role") ?? "standard"
-        let role: ConfirmContent.ConfirmRole = (roleStr == "destructive") ? .destructive : .standard
-        // Cap user-supplied timeout at 300s so a malformed or malicious caller
-        // can't starve the socket worker pool with a .distantFuture wait.
-        let maxTimeoutSeconds: Double = 300
-        let timeoutSeconds = (params["timeout"] as? Double).map { min(maxTimeoutSeconds, max(0, $0)) }
-        let clientId = (v2String(params, "_clientId")) ?? "socket"
-
-        // Optional confirm/cancel label overrides. Falls back to a neutral
-        // "OK" key (dialog.pane.confirm.ok) — NOT the close-specific key that
-        // localizes to "Close"/"閉じる" and is semantically wrong for generic
-        // socket-initiated prompts (synthesis-standard §1.3).
-        let confirmLabel = v2String(params, "confirm_label").flatMap { $0.isEmpty ? nil : $0 }
-            ?? String(localized: "dialog.pane.confirm.ok", defaultValue: "OK")
-        let cancelLabel = v2String(params, "cancel_label").flatMap { $0.isEmpty ? nil : $0 }
-            ?? String(localized: "dialog.pane.confirm.cancel", defaultValue: "Cancel")
-
-        let semaphore = DispatchSemaphore(value: 0)
-        // `outcome` is written from the main-actor completion callback and read
-        // after the semaphore is signaled — single-writer, single-reader with
-        // happens-before ordering provided by the semaphore.
-        final class OutcomeHolder {
-            var value: ConfirmResult = .dismissed
-            var fired: Bool = false
-        }
-        let holder = OutcomeHolder()
-        var presented = false
-        var presentedInteractionId: UUID?
-
-        v2MainSync {
-            // Find the workspace that owns this panel. Iterate tabs across the
-            // resolved TabManager — pane.confirm accepts any workspace's panel.
-            guard let workspace = tabManager.tabs.first(where: { $0.panels[panelId] != nil }) else {
-                return
-            }
-            let content = ConfirmContent(
-                title: title,
-                message: message.isEmpty ? nil : message,
-                confirmLabel: confirmLabel,
-                cancelLabel: cancelLabel,
-                role: role,
-                source: .socket(clientId: clientId),
-                completion: { result in
-                    holder.value = result
-                    holder.fired = true
-                    semaphore.signal()
-                }
-            )
-            workspace.paneInteractionRuntime.present(
-                panelId: panelId,
-                interaction: .confirm(content)
-            )
-            presentedInteractionId = content.id
-            presented = true
-        }
-
-        guard presented else {
-            return .err(code: "unknown_panel", message: "Panel not found", data: ["panel_id": panelId.uuidString])
-        }
-
-        // Block the socket thread until the user responds or the timeout fires.
-        let waitDeadline: DispatchTime
-        if let timeoutSeconds {
-            waitDeadline = .now() + timeoutSeconds
-        } else {
-            waitDeadline = .now() + maxTimeoutSeconds
-        }
-        let waitResult = semaphore.wait(timeout: waitDeadline)
-        if waitResult == .timedOut {
-            // Accept/timeout race: the user may have clicked Confirm just as the
-            // timeout fired. Re-read holder on the main thread before cancelling,
-            // and if the completion fired (holder.fired == true) return the user's
-            // actual outcome instead of overwriting with dismissed.
-            let raced: ConfirmResult? = v2MainSync {
-                if holder.fired {
-                    return holder.value
-                }
-                if let workspace = tabManager.tabs.first(where: { $0.panels[panelId] != nil }) {
-                    // Interaction-ID-guarded cancel: never cancel a successor
-                    // that advanced after our present (synthesis-critical §1.5).
-                    workspace.paneInteractionRuntime.cancelActive(
-                        panelId: panelId,
-                        ifInteractionId: presentedInteractionId
-                    )
-                }
-                return nil
-            }
-            if let raced {
-                return v2EncodePaneConfirmResult(raced)
-            }
-            return .ok(["result": "dismissed"])
-        }
-
-        return v2EncodePaneConfirmResult(holder.value)
-    }
-
-    private func v2EncodePaneConfirmResult(_ result: ConfirmResult) -> V2CallResult {
-        switch result {
-        case .confirmed:
-            return .ok(["result": "ok"])
-        case .cancelled:
-            return .ok(["result": "cancel"])
-        case .dismissed:
-            return .ok(["result": "dismissed"])
-        }
-    }
 
     // MARK: - V2 Conversation Methods (C11-24)
